+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - new property "theme": defines a set of customizations applied over the selected **Style**. In other words, a stylesheet file. At the moment 2 are available:
         - Default - Light: icons and colors for light desktop environments
         - Darcula: dark theme based on JetBrain's Darcula theme
+    - "theme" property is changed through a new lower bar button:
+    <p align="center">
+        <img src="https://raw.githubusercontent.com/vinifmor/bauh-files/master/pictures/releases/0.9.9/bt_themes.png">
+    </p>
         
 ### Improvements
 - UI